+++ config/ember-try.js
@@ -30,6 +30,7 @@ module.exports = function() {
           npm: {
             devDependencies: {
               'ember-source': '~3.4.0',
+              'ember-native-class-polyfill': '~1.0.6',
             },
           },
         },