+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
 
 - `graphql` has been added as a peer dependency and removed from the addon's direct dependency. Please run `ember install ember-apollo-client` to add the default dependencies to your project or run `yarn add -D graphql`.
+- If you are using Ember versions 3.4 or 3.5 you must add `ember-native-class-polyfil` to your application. Please run `ember install ember-native-class-polyfil`. Later Ember versions are not required to have this polyfil.
 
 
 ## [v2.0.0-beta.3] - 2019-03-09